Adds FileManagerFragment
